+++ projects/pip-webui2-behaviors/src/lib/hotkeys/hotkeys.module.ts
@@ -1,17 +1,9 @@
-import { NgModule, ModuleWithProviders } from '@angular/core';
+import { NgModule } from '@angular/core';
 
 import { PipHotkeysService } from './shared/hotkeys.service';
 
 @NgModule({
     declarations: [],
+    providers: [PipHotkeysService]
 })
-export class PipHotkeysModule {
-    static forRoot(): ModuleWithProviders {
-        return {
-            ngModule: PipHotkeysModule,
-            providers: [
-                PipHotkeysService
-            ]
-        };
-    }
-}
+export class PipHotkeysModule { }