+++ projects/pip-webui2-behaviors/src/lib/hotkeys/shared/index.ts
@@ -1,1 +1,2 @@
 export * from './hotkeys.service';
+export * from './models';