+++ projects/pip-webui2-behaviors/src/lib/shortcuts/index.ts
@@ -1,2 +1,0 @@
-export * from './shared/index';
-export { PipShortcutsModule as PipShortcutModule } from './shortcut.module';