+++ projects/pip-webui2-behaviors/src/lib/shortcuts/shared/index.ts
@@ -1,6 +1,0 @@
-export * from './Shortcut';
-export * from './ShortcutItem';
-export * from './ShortcutOptions';
-export * from './ShortcutsConfig';
-export * from './shortcut.binding.service';
-export * from './shortcut.service';