+++ src/app/app-routing.module.ts
@@ -6,7 +6,6 @@ import { FocusedExampleComponent } from './focused-example/focused-example.compo
 import { InfiniteScrollExampleComponent } from './infinite-scroll-example/infinite-scroll-example.component';
 import { InfiniteScrollPageExampleComponent } from './infinite-scroll-page-example/infinite-scroll-page-example.component';
 import { SelectedExampleComponent } from './selected-example/selected-example.component';
-import { ShortcutsExampleComponent } from './shortcuts-example/shortcuts-example.component';
 import { HotkeysExampleComponent } from './hotkeys-example/hotkeys-example.component';
 
 const appRoutes: Routes = [
@@ -15,7 +14,6 @@ const appRoutes: Routes = [
     { path: 'infinite_scroll', component: InfiniteScrollExampleComponent },
     { path: 'infinite_page', component: InfiniteScrollPageExampleComponent },
     { path: 'draggable', component: DraggableExampleComponent },
-    { path: 'shortcuts', component: ShortcutsExampleComponent },
     { path: 'hotkeys', component: HotkeysExampleComponent },
     { path: '**', redirectTo: 'selected' },
 ];