+++ src/app/hotkeys-example/hotkeys-example.component.ts
@@ -1,8 +1,6 @@
 import { Component, Injector, SkipSelf } from '@angular/core';
 import { PipHotkeysService } from 'pip-webui2-behaviors';
 
-import { HotkeysExampleModule } from './hotkeys-example.module';
-
 @Component({
     selector: 'app-hotkeys-example',
     templateUrl: './hotkeys-example.component.html',
@@ -11,27 +9,30 @@ import { HotkeysExampleModule } from './hotkeys-example.module';
 })
 export class HotkeysExampleComponent {
 
+    public raisedBg = false;
+    public raisedFg = false;
+
     constructor(
         private pipHotkeys: PipHotkeysService,
         @SkipSelf() private injector: Injector
     ) {
         this.pipHotkeys.add({
             hotkey: 'alt+ h',
-            action: () => { alert('alt+h pressed'); }
+            action: () => { this.raisedBg = !this.raisedBg; }
         });
         this.pipHotkeys.add({
             hotkey: 'alt+h',
-            action: () => { alert('shouldnt exist'); }
+            action: () => { console.log('shouldnt exist'); }
         });
         this.pipHotkeys.add({
             hotkey: 'alt+.',
-            action: () => { alert('Local alt+.'); }
+            action: () => { this.raisedFg = !this.raisedFg; }
           });
         const globalPipHotkeys: PipHotkeysService = this.injector.get(PipHotkeysService);
-        globalPipHotkeys.remove('alt+g');
+        globalPipHotkeys.remove('ctrl+alt+f');
         globalPipHotkeys.add({
-            hotkey: 'alt+g',
-            action: () => { alert('New global alt+g'); }
+            hotkey: 'ctrl+alt+f',
+            navigationCommand: ['selected']
         });
     }